moved settings functions to new lib and updated all imports
