+++ lib/ads.js
@@ -1,4 +1,4 @@
-export function getArticleAds() {
+export function getArticleAds(params) {
   let today = new Date();
   const dd = String(today.getDate()).padStart(2, '0');
   const mm = String(today.getMonth() + 1).padStart(2, '0');
@@ -6,10 +6,10 @@ export function getArticleAds() {
 
   today = `${yyyy}-${mm}-${dd}`;
 
-  return fetch(`${process.env.LETTERHEAD_API_URL}promotions?date=${today}`, {
+  return fetch(`${params['url']}promotions?date=${today}`, {
     headers: {
       'Content-Type': 'application/json',
-      Authorization: `Bearer ${process.env.LETTERHEAD_API_KEY}`,
+      Authorization: `Bearer ${params['apiKey']}`,
     },
   })
     .then((response) => response.json())
