+++ lib/articles.js
@@ -2610,25 +2610,3 @@ export async function generateAllArticleArchivePages(params = {}) {
 
   return allPaths;
 }
-
-const LIST_SETTINGS = `query FrontendGetOrgSettings {
-  settings {
-    name
-    value
-  }
-  organization_locales {
-    locale {
-      code
-    }
-  }
-}`;
-
-export async function getOrgSettings(params) {
-  return fetchGraphQL({
-    url: params['url'],
-    adminSecret: process.env.HASURA_ADMIN_SECRET,
-    site: params['site'],
-    query: LIST_SETTINGS,
-    name: 'FrontendGetOrgSettings',
-  });
-}
