+++ lib/utils.js
@@ -606,20 +606,3 @@ export const parseDonorViews = (queryResult, limit) => {
   });
   return dv;
 };
-
-export function findSetting(settings, name) {
-  try {
-    let item = settings.find((setting) => setting.name === name);
-    return item.value;
-  } catch (err) {
-    console.error(`Error finding setting ${name}:`, err);
-  }
-}
-
-export function booleanSetting(items, name, defaultValue) {
-  let setting = items.find((item) => item.name === name);
-  if (setting) {
-    return setting.value;
-  }
-  return defaultValue;
-}
