+++ pages/_sites/[site]/newsletters/index.js
@@ -4,11 +4,12 @@ import { cachedContents } from '../../../../lib/cached';
 import { hasuraListNewsletters } from '../../../../lib/newsletters.js';
 import { getArticleAds } from '../../../../lib/ads.js';
 import ArticleStream from '../../../../components/homepage/ArticleStream';
+import { generateAllDomainPaths } from '../../../../lib/articles.js';
 import {
-  generateAllDomainPaths,
+  booleanSetting,
+  findSetting,
   getOrgSettings,
-} from '../../../../lib/articles.js';
-import { booleanSetting } from '../../../../lib/utils.js';
+} from '../../../../lib/settings.js';
 
 export default function NewsletterIndexPage(props) {
   const isAmp = false;
@@ -101,14 +102,16 @@ export async function getStaticProps({ params }) {
     }
   }
 
+  const settings = settingsResult.data.settings;
   let expandedAds = [];
-  let letterheadSetting = booleanSetting(
-    settingsResult.data.settings,
-    'LETTERHEAD_API_URL',
-    false
-  );
+  let letterheadSetting = booleanSetting(settings, 'LETTERHEAD_API_URL', false);
   if (letterheadSetting) {
-    const allAds = (await cachedContents('ads', getArticleAds)) || [];
+    let letterheadParams = {
+      url: findSetting(settings, 'LETTERHEAD_API_URL'),
+      apiKey: findSetting(settings, 'LETTERHEAD_API_KEY'),
+    };
+    const allAds =
+      (await cachedContents('ads', getArticleAds(letterheadParams))) || [];
     expandedAds = allAds.filter((ad) => ad.adTypeId === 166 && ad.status === 4);
   }
