+++ pages/_sites/[site]/preview/[category]/[slug].js
@@ -6,6 +6,11 @@ import {
 import { getArticleAds } from '../../../../../lib/ads.js';
 import { cachedContents } from '../../../../../lib/cached';
 import Article from '../../../../../components/Article.js';
+import {
+  booleanSetting,
+  findSetting,
+  getOrgSettings,
+} from '../../../../../lib/settings.js';
 
 export default function PreviewArticle(props) {
   if (!props.article) {
@@ -44,6 +49,16 @@ export async function getStaticProps(context) {
       notFound: true,
     };
   }
+
+  const settingsResult = await getOrgSettings({
+    url: apiUrl,
+    site: site,
+  });
+
+  if (settingsResult.errors) {
+    throw settingsResult.errors;
+  }
+  const settings = settingsResult.data.settings;
 
   let article = {};
   let sectionArticles = [];
@@ -99,8 +114,14 @@ export async function getStaticProps(context) {
   }
 
   let ads = [];
-  if (process.env.LETTERHEAD_API_URL) {
-    const allAds = (await cachedContents('ads', getArticleAds)) || [];
+  let letterheadSetting = booleanSetting(settings, 'LETTERHEAD_API_URL', false);
+  if (letterheadSetting) {
+    let letterheadParams = {
+      url: findSetting(settings, 'LETTERHEAD_API_URL'),
+      apiKey: findSetting(settings, 'LETTERHEAD_API_KEY'),
+    };
+    const allAds =
+      (await cachedContents('ads', getArticleAds(letterheadParams))) || [];
     ads = allAds.filter((ad) => ad.adTypeId === 164 && ad.status === 4);
   }
