+++ pages/_sites/[site]/tinycms/homepage-layouts/index.js
@@ -4,8 +4,7 @@ import { useRouter } from 'next/router';
 import { hasuraListHomepageLayoutSchemas } from '../../../../../lib/homepage.js';
 import AdminLayout from '../../../../../components/AdminLayout.js';
 import AdminNav from '../../../../../components/nav/AdminNav';
-import { getOrgSettings } from '../../../../../lib/articles.js';
-import { findSetting } from '../../../../../lib/utils';
+import { findSetting, getOrgSettings } from '../../../../../lib/settings.js';
 
 export default function HomepageLayouts({ homepageLayouts, siteUrl, host }) {
   const [message, setMessage] = useState(null);
