+++ pages/api/sidebar/documents/[id].js
@@ -1,10 +1,9 @@
 import {
-  getOrgSettings,
   hasuraLookupGoogleDoc,
   hasuraGetGoogleDocsForArticle,
   hasuraGetGoogleDocsForPage,
 } from '../../../../lib/articles';
-import { findSetting } from '../../../../lib/utils.js';
+import { findSetting, getOrgSettings } from '../../../../lib/settings.js';
 
 export default async function Handler(req, res) {
   const apiUrl = process.env.HASURA_API_URL;
