+++ pages/api/sidebar/documents/[id]/preview.js
@@ -1,8 +1,5 @@
-import {
-  getOrgSettings,
-  hasuraLookupGoogleDoc,
-} from '../../../../../lib/articles';
-import { findSetting } from '../../../../../lib/utils.js';
+import { hasuraLookupGoogleDoc } from '../../../../../lib/articles';
+import { findSetting, getOrgSettings } from '../../../../../lib/settings';
 import {
   processDocumentContents,
   saveArticle,
